# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen Kursus & Perencanaan API",
        "description": "L5 Swagger API documentation for Manajemen Kursus",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quizzes/{quiz}": {
            "get": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Pembuka kuis (informasi terkait pengerjaan kuis & berisi riwayat pengerjaan kuis karna bisa saja remidi)",
                "description": "Menampilkan informasi sebelum memulai kuis beserta riwayatnya",
                "operationId": "quizInformation",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informasi kuis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quiz information"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Pengerjaan kuis",
                "description": "Mengirimkan jawaban kuis",
                "operationId": "submitQuiz",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "answers"
                                ],
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "string",
                                                    "example": "uuid-question"
                                                },
                                                "answer_id": {
                                                    "type": "string",
                                                    "example": "uuid-answer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kuis berhasil disubmit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Submit quiz"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/quizzes/result/{quiz}": {
            "get": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Detail hasil pengerjaan kuis",
                "description": "Melihat hasil pengerjaan kuis",
                "operationId": "quizResult",
                "parameters": [
                    {
                        "name": "quiz",
                        "in": "path",
                        "description": "Quiz ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil kuis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quiz result"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}